java_input_output_Object_append

+++ 국비지원학원/java/22_input_output_stream/src/f10_object/append/ObjectOutputExample.java
@@ -8,9 +8,18 @@ public class ObjectOutputExample {
 
 	public static void main(String[] args) throws Exception{
 		File file = new File("person.dat");
-		ObjectOutputStream oos = new ObjectOutputStream(
-						new FileOutputStream(file,true)
-					);
+		ObjectOutputStream oos = null;
+		// 파일이 존재하거나 파일의 크기가 0보다 크면 기존의 데이터가 존재
+		if(file.exists() || file.length() > 0) {
+			oos = new ObjectOutputAppend(
+					new FileOutputStream(file,true)
+				);
+		}else {
+			oos = new ObjectOutputStream(
+					new FileOutputStream(file,true)
+				);
+		}
+				
 		Player p = new Player();
 		p.setpNum(1);
 		p.setNick("신");
